refactor(ui): share router lifecycle for reconnect recovery
