+++ ui/src/app-routes.ts
@@ -21,6 +21,7 @@ import {
   INTERNAL_TERMINAL_PATH_PARAM,
   INTERNAL_WORKBOARD_PATH_PARAM,
   isLegacyPluginsDiscoveryPath,
+  isSessionRouteId,
   memoryTabFromPath,
   pathForAgentPanel,
   pathForRoute,
@@ -36,7 +37,7 @@ import {
   type RouteId,
 } from "./app-route-paths.ts";
 import type { ApplicationContext } from "./app/context.ts";
-import { startSessionRouteRecovery } from "./app/session-route-recovery.ts";
+import { gatewayPresentationScope } from "./app/gateway-presentation-scope.ts";
 import { page as aboutPage } from "./pages/about/route.ts";
 import { page as activityPage } from "./pages/activity/route.ts";
 import { page as agentsPage } from "./pages/agents/route.ts";
@@ -303,9 +304,75 @@ export async function startApplicationRouter(
     replace: (next) => history.replace(next),
     listen: (listener) => {
       let listening = true;
-      const stopSessionRecovery = startSessionRouteRecovery(router, context);
+      let recoveryQueued = false;
+      let interrupted:
+        | { controller: AbortController; scope: ReturnType<typeof gatewayPresentationScope> }
+        | undefined;
+      const currentTarget = () => {
+        const state = router.getState();
+        return state.pendingMatches[0] ?? state.matches[0];
+      };
+      const recoverSessionRoute = () => {
+        const target = currentTarget();
+        if (!target || !isSessionRouteId(target.routeId)) {
+          interrupted = undefined;
+          return;
+        }
+        const scope = gatewayPresentationScope(context.gateway);
+        if (interrupted?.controller !== target.abortController) {
+          interrupted = undefined;
+        }
+        if (interrupted && interrupted.scope !== scope) {
+          return;
+        }
+        if (context.gateway.snapshot.phase !== "connected") {
+          if (target.status === "pending" || target.isFetching === "loader") {
+            interrupted = { controller: target.abortController, scope };
+          }
+          return;
+        }
+        if (target.status === "success" && !target.isFetching) {
+          interrupted = undefined;
+        }
+        if (!interrupted || recoveryQueued || target.status !== "error") {
+          return;
+        }
+        recoveryQueued = true;
+        // Other subscribers may navigate synchronously; recover only their final intent.
+        queueMicrotask(() => {
+          recoveryQueued = false;
+          const latest = currentTarget();
+          if (
+            !listening ||
+            !interrupted ||
+            latest?.abortController !== interrupted.controller ||
+            gatewayPresentationScope(context.gateway) !== interrupted.scope ||
+            context.gateway.snapshot.phase !== "connected" ||
+            latest.status !== "error"
+          ) {
+            return;
+          }
+          interrupted = undefined;
+          // The loader publishes its error before retiring its run. Abort it so
+          // same-match revalidation cannot join the already failed promise.
+          latest.abortController.abort();
+          if (currentTarget()?.abortController !== latest.abortController) {
+            return;
+          }
+          void router
+            .navigate(
+              latest.routeId,
+              context,
+              { history: "none", revalidate: true },
+              latest.location,
+            )
+            .catch(() => undefined);
+        });
+      };
+      const stopSessionRecovery = router.subscribe(recoverSessionRoute);
       let lastHello = context.gateway.snapshot.hello;
       const stopGateway = context.gateway.subscribe((snapshot) => {
+        recoverSessionRoute();
         if (lastHello === snapshot.hello) {
           return;
         }
@@ -360,6 +427,7 @@ export async function startApplicationRouter(
       });
       return () => {
         listening = false;
+        interrupted = undefined;
         stopSessionRecovery();
         stopGateway();
         stopHistory();
